+++ app/listTime.tsx
@@ -1,4 +1,4 @@
-import { View, Text, StyleSheet, ScrollView } from "react-native";
+import { View, StyleSheet, ScrollView } from "react-native";
 import React from "react";
 import TimeIcon from "../components/icons/TimeIcon";
 import TitleApp from "../components/title/TitleApp";
