+++ components/forms/FormLogin.tsx
@@ -1,14 +1,11 @@
-import { View, Text, StyleSheet, TouchableOpacity } from "react-native";
+import { View, Text, StyleSheet } from "react-native";
 import React from "react";
 import { useFonts, OpenSans_300Light } from "@expo-google-fonts/open-sans";
-import Input from "../inputs/Input";
-import { useNavigation, useRouter } from "expo-router";
+import Input from "../inputs/InputForms";
 import Button from "../buttons/Button";
 
 export default function FormLogin() {
-  const router = useRouter();
   const redirectLogin = () => {
-    router.push("/calculateHours");
     console.log("redirecionado para calculateHours");
   };
 
@@ -20,22 +17,15 @@ export default function FormLogin() {
 
   return (
     <View style={styles.container}>
-      <Input
-        labelStyle={styles.label}
-        inputStyle={styles.input}
-        label="User"
-        placeholder="Digite seu email"
-      />
+      <Input label="User" placeholder="Digite seu email" />
 
       <Input
-        labelStyle={styles.label}
-        inputStyle={styles.input}
         label="Password"
         placeholder="Digite sua senha"
         secureTextEntry={true}
       />
 
-      <Button buttonName="Entrar" redirect={redirectLogin} />
+      <Button buttonName="Entrar" submitFunction={redirectLogin} />
 
       <Text style={styles.textLink}>
         You don't have an account?{" "}
@@ -53,20 +43,6 @@ export default function FormLogin() {
 const styles = StyleSheet.create({
   container: {
     alignItems: "center",
-  },
-  input: {
-    width: 350,
-    backgroundColor: "#fff",
-    borderRadius: 10,
-    marginTop: 5,
-    paddingLeft: 10,
-    marginBottom: 20,
-  },
-  label: {
-    color: "#fff",
-    fontSize: 20,
-    textAlign: "left",
-    fontFamily: "OpenSans_300Light",
   },
   textLink: {
     color: "#fff",
